Remove Linux System tray support Closes #996
