+++ src/qz/ui/tray/TaskbarTrayIcon.java
@@ -1,6 +1,7 @@
 package qz.ui.tray;
 
 import qz.common.Constants;
+import qz.utils.LinuxUtilities;
 import qz.utils.UbuntuUtilities;
 import qz.utils.UnixUtilities;
 
@@ -27,11 +28,6 @@ public class TaskbarTrayIcon extends JFrame implements WindowListener {
         setTaskBarTitle(getTitle());
         setSize(0, 0);
         getContentPane().setBackground(Color.BLACK);
-        if (UnixUtilities.isUbuntu()) {
-            // attempt to camouflage the single pixel left behind
-            getContentPane().setBackground(UbuntuUtilities.getTrayColor());
-        }
-
         iconSize = new Dimension(40, 40);
 
         setIconImage(trayImage);
